Exclude self from static cluster connectors
The static-connectors of a cluster should not contain this brokers'
own connector as it will happily connect to itself and thus not
properly participate in the cluster.

Also added an option to configure allow-direct-connections-only and
set it to false. This would allow a cluster to become dynamic using
an initial static list.